feat: improve mobile proposal item styles
