+++ src/components/Layout/ActionableLayout.css
@@ -4,7 +4,12 @@
   flex-wrap: wrap;
   justify-content: space-between;
   align-items: center;
-  margin-bottom: 16px;
+}
+
+@media (min-width: 768px) {
+  .ActionableLayout__Action {
+    margin-bottom: 16px;
+  }
 }
 
 .ActionableLayout__Right {
