+++ src/components/Search/SortingMenu.css
@@ -1,6 +1,10 @@
 .SortingMenu {
   margin-right: 1em;
   font-size: 12px !important;
+  height: 40px;
+  display: flex !important;
+  justify-content: center;
+  align-items: center;
 }
 
 .SortingMenu .divider.text {
