--- src/components/Status/FinishLabel.css
@@ -1,16 +1,0 @@
-.FinishLabel {
-  padding: 0;
-  height: 24px;
-  display: inline-flex;
-  color: var(--black-800);
-  font-weight: var(--weight-semi-bold);
-  font-size: 12px;
-  line-height: 26px;
-  vertical-align: middle;
-  align-content: center;
-  align-items: center;
-  text-transform: uppercase;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
