+++ src/ui-overrides.css
@@ -226,3 +226,7 @@
   margin: 0 !important;
   padding: 0 !important;
 }
+
+.dcl.login-modal .dcl.option {
+  width: 100% !important;
+}
